So if models with types we keep together. Let move it all to schema folder and make small refactoring.
