+++ server/index.js
@@ -2,29 +2,25 @@ const Koa = require('koa');
 const bodyParser = require('koa-bodyparser');
 const { graphqlKoa, graphiqlKoa } = require('graphql-server-koa');
 const Router = require('koa-router');
+const schema = require('./schema');
+
+const PORT = '3000';
 
 const router = new Router();
-
-const schema = require('./schema.js');
-
-
-console.log(schema);
 const app = new Koa();
 
-const PORT = '3000'
-
-router.post('/graphql', graphqlKoa(ctx => ({
-  schema,
-})));
+router.post(
+  '/graphql',
+  graphqlKoa(ctx => ({
+    schema,
+  }))
+);
 
 router.get('/graphiql', graphiqlKoa({ endpointURL: '/graphql' }));
 
-app
-  .use(bodyParser())
-  .use(router.routes())
-;
+app.use(bodyParser()).use(router.routes());
 
-app.on('error', (err) => {
+app.on('error', err => {
   console.error('server error', err);
 });
