--- server/schema.js
@@ -1,5 +1,0 @@
-const { makeExecutableSchema, mergeSchemas } = require('graphql-tools');
-
-const { GQCType } = require('./types');
-
-module.exports = GQCType;
