+++ server/schema/Appointment.js
