--- server/types/index.js
@@ -1,5 +1,0 @@
-const GQCType = require('./GQC');
-
-module.exports = {
-  GQCType
-};
